Add files via upload
The reason ageism exists in a company is that workers refuse to embrace new ideas, techniques and advance their education.  With the world evolving, employees and employers need to advance knowledge, to be more competitive in the world.  In order to, reduce ageism, employees need to change their mindset and companies need to embrace education programs because older employees will stay at a company, than younger employees.   Employees at willing to leave at the 5-year mark.  This could be because of more advancement opportunities dealing with education, new skills and/or job promotions.  Company and employees’ relations need to improve in order for an employee to stay at a company.  Companies need to establish a lifestyle environment, where employees feel part of the company. Employees want to know that they matter and the company is listening to them.   Income is important but feeling part of something is more rewarding. Laying-off or not hiring older employees is not the answer.  Education and the environment are important to make employees stay.  When a company invests in its employees, employees will exceed and grow. Using a metric for education in an employee's review will force an employee to obtain more knowledge and will help the company be more competitive.  Employees need to change the mindset of advancing themselves for the greater good.